microservice1/requirements.txt
Flask
pandas
psycopg2-binary
requests
